+++ src/Icons/FontAwesome6IconTheme.php
@@ -3,12 +3,12 @@
 declare(strict_types=1);
 
 /**
- * Contains the FontAwesomeIconTheme class.
+ * Contains the FontAwesome6IconTheme class.
  *
- * @copyright   Copyright (c) 2020 Attila Fulop
+ * @copyright   Copyright (c) 2022 Attila Fulop
  * @author      Attila Fulop
  * @license     MIT
- * @since       2020-10-18
+ * @since       2022-02-22
  *
  */
 
@@ -17,68 +17,68 @@ namespace Konekt\AppShell\Icons;
 use Konekt\AppShell\Contracts\IconTheme;
 use Konekt\AppShell\Theme\ThemeColor;
 
-class FontAwesomeIconTheme implements IconTheme
+class FontAwesome6IconTheme implements IconTheme
 {
     use HasIconMap;
     use HasFallbackIcon;
 
-    public const ID = 'font-awesome';
+    public const ID = 'font-awesome6';
 
-    private const FA_ANIMATION_TYPES = ['pulse', 'spin'];
+    private const FA_ANIMATION_TYPES = ['beat', 'beat-fade', 'bounce', 'flip', 'shake', 'spin'];
 
     private static string $fallbackIcon = 'bookmark';
 
     private static array $icons = [
-        AppShellIcons::USERS => 'user-friends',
+        AppShellIcons::USERS => 'user-group',
         AppShellIcons::USER => 'user',
-        AppShellIcons::USER_ACTIVE => 'user',
-        AppShellIcons::USER_INACTIVE => 'user-slash',
+        AppShellIcons::USER_ACTIVE => 'user-check',
+        AppShellIcons::USER_INACTIVE => 'user-xmark',
         AppShellIcons::CUSTOMERS => 'address-card',
         AppShellIcons::CUSTOMER => 'id-badge',
-        AppShellIcons::WARNING => 'exclamation-triangle',
-        AppShellIcons::ACTIVE => 'chevron-circle-right',
-        AppShellIcons::INACTIVE => 'times-circle',
+        AppShellIcons::WARNING => 'triangle-exclamation',
+        AppShellIcons::ACTIVE => 'circle-chevron-right',
+        AppShellIcons::INACTIVE => 'circle-xmark',
         AppShellIcons::CHECK => 'check',
-        AppShellIcons::CROSS => 'times',
-        AppShellIcons::INFO => 'info-circle',
-        AppShellIcons::HELP => 'question-circle',
+        AppShellIcons::CROSS => 'xmark',
+        AppShellIcons::INFO => 'circle-info',
+        AppShellIcons::HELP => 'circle-question',
         AppShellIcons::EMAIL => 'envelope',
-        AppShellIcons::ORGANIZATION => 'building',
+        AppShellIcons::ORGANIZATION => 'city',
         AppShellIcons::PASSWORD => 'key',
-        AppShellIcons::MONEY => 'money-check-alt',
-        AppShellIcons::CALENDAR => 'calendar-alt',
-        AppShellIcons::MALE => 'male',
-        AppShellIcons::FEMALE => 'female',
-        AppShellIcons::SMILEY_GLAD => 'smile',
-        AppShellIcons::SMILEY_SAD => 'sad-tear',
-        AppShellIcons::MORE_ITEMS => 'ellipsis-v',
+        AppShellIcons::MONEY => 'money-check-dollar',
+        AppShellIcons::CALENDAR => 'calendar-day',
+        AppShellIcons::MALE => 'person',
+        AppShellIcons::FEMALE => 'person-dress',
+        AppShellIcons::SMILEY_GLAD => 'face-smile',
+        AppShellIcons::SMILEY_SAD => 'face-frown',
+        AppShellIcons::MORE_ITEMS => 'ellipsis-vertical',
         AppShellIcons::PLUS => 'plus',
         AppShellIcons::MINUS => 'minus',
         AppShellIcons::EDIT => 'pen',
-        AppShellIcons::DELETE => 'times',
+        AppShellIcons::DELETE => 'xmark',
         AppShellIcons::TAG => 'tag',
         AppShellIcons::COLOR => 'palette',
         AppShellIcons::SORT => 'sort',
-        AppShellIcons::SORT_ASC => 'sort-amount-up',
-        AppShellIcons::SORT_DESC => 'sort-amount-down',
-        AppShellIcons::SETTINGS => 'cog',
-        AppShellIcons::SECURITY => 'shield-alt',
+        AppShellIcons::SORT_ASC => 'arrow-up-wide-short',
+        AppShellIcons::SORT_DESC => 'arrow-down-wide-short',
+        AppShellIcons::SETTINGS => 'gear',
+        AppShellIcons::SECURITY => 'shield',
         AppShellIcons::STAR => 'star',
-        AppShellIcons::TIME => 'history',
+        AppShellIcons::TIME => 'clock-rotate-left',
         AppShellIcons::TAX => 'scroll',
         AppShellIcons::QUICK_LINKS => 'bolt',
-        AppShellIcons::HAMBURGER => 'bars',
+        AppShellIcons::HAMBURGER => 'bars-staggered',
         AppShellIcons::LINK => 'link',
         AppShellIcons::IMAGE => 'images',
         AppShellIcons::CHEVRON_RIGHT => 'chevron-right',
         AppShellIcons::CHEVRON_LEFT => 'chevron-left',
-        AppShellIcons::FILTERS => 'sliders-h',
-        AppShellIcons::SEARCH => 'search',
+        AppShellIcons::FILTERS => 'sliders',
+        AppShellIcons::SEARCH => 'magnifying-glass',
     ];
 
     public static function getName(): string
     {
-        return 'Font Awesome 5 Icons';
+        return 'Font Awesome 6 Icons';
     }
 
     public function get(string $abstract): string
@@ -92,7 +92,7 @@ class FontAwesomeIconTheme implements IconTheme
             return '';
         }
 
-        return '<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.1/css/all.min.css" integrity="sha512-+4zCK9k+qNFUR5X+cKL9EIR+ZOhtIloNl9GIKS57V1MyNsYpYcUrUeQc9vNfzsWfV28IaLL3i96P9sdNyeRssA==" crossorigin="anonymous" />';
+        return '<link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css" integrity="sha512-9usAa10IRO0HhonpyAIVpjrylPvoDwiPUiKdWk5t3PyolY1cOd4DSE0Ga+ri4AuTroPR5aQvXU9xC6qOPnzFeg==" crossorigin="anonymous" referrerpolicy="no-referrer" />';
     }
 
     public function render(string $abstract, ThemeColor $color, array $attributes = []): string
@@ -125,7 +125,7 @@ class FontAwesomeIconTheme implements IconTheme
         );
 
         return sprintf(
-            '<i class="fas fa-%s %s" %s></i>',
+            '<i class="fa-solid fa-%s %s" %s></i>',
             $this->get($abstract),
             implode(' ', $classes),
             $attrString